requirements.txt
attrs==21.4.0
iniconfig==1.1.1
lxml==4.7.1
packaging==21.3
Pillow==9.0.0
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.6
pytest==6.2.5
python-pptx==0.6.21
toml==0.10.2
XlsxWriter==3.0.2
